pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.4",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rasa-sdk"
version = "3.8.0"
description = "Open source machine learning framework to automate text- and voice-based conversations: NLU, dialogue management, connect to Slack, Facebook, and more - Create chatbots and voice assistants"
authors = [ "Rasa Technologies GmbH <hi@rasa.com>",]
maintainers = [ "Tom Bocklisch <tom@rasa.com>",]
homepage = "https://rasa.com"
repository = "https://github.com/rasahq/rasa-sdk"
documentation = "https://rasa.com/docs"
classifiers = [ "Development Status :: 5 - Production/Stable", "Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Topic :: Software Development :: Libraries",]
keywords = [ "nlp", "machine-learning", "machine-learning-library", "bot", "bots", "botkit", "rasa conversational-agents", "conversational-ai", "chatbot", "chatbot-framework", "bot-framework",]
include = [ "LICENSE.txt", "README.md",]
readme = "README.md"
license = "Apache-2.0"

[tool.towncrier]
package = "rasa_sdk"
package_dir = "rasa_sdk"
filename = "CHANGELOG.mdx"
directory = "./changelog"
underlines = "   "
title_format = "## [{version}] - {project_date}"
template = "./changelog/_template.md.jinja2"
start_string = "<!-- TOWNCRIER -->\n"
issue_format = "[#{issue}](https://github.com/rasahq/rasa-sdk/issues/{issue})"
[[tool.towncrier.type]]
directory = "removal"
name = "Deprecations and Removals"
showcontent = true

[[tool.towncrier.type]]
directory = "feature"
name = "Features"
showcontent = true

[[tool.towncrier.type]]
directory = "improvement"
name = "Improvements"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bugfixes"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Improved Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "misc"
name = "Miscellaneous internal changes"
showcontent = false

[tool.mypy]
ignore_missing_imports = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.ruff]
line-length = 88
target-version = "py38"
exclude = [ "rasa_sdk/grpc_py", "eggs", ".git", ".pytest_cache", "build", "dist", ".DS_Store"]

[tool.ruff.lint]
ignore = [ "D100", "D101", "D102", "D103", "D104", "D105", "RUF005",]
select = [ "D", "E", "F", "W", "RUF",]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
coloredlogs = ">=10,<16"
sanic = "^21.12.0"
typing-extensions = ">=4.1.1,<5.0.0"
Sanic-Cors = "^2.0.0"
prompt-toolkit = "^3.0,<3.0.29"
"ruamel.yaml" = ">=0.16.5,<0.18.0"
websockets = ">=10.0,<12.0"
pluggy = "^1.0.0"
opentelemetry-api = "~1.15.0"
opentelemetry-sdk = "~1.15.0"
opentelemetry-exporter-jaeger = "~1.15.0"
opentelemetry-exporter-otlp = "~1.15.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^4.1.0"
coveralls = "^3.0.1"
pytest = "^7.4.0"
black = "22.12.0"
questionary = ">=1.5.1,<2.1.0"
towncrier = "^23.6.0"
toml = "^0.10.0"
pep440-version-utils = "^0.3.0"
semantic_version = "^2.8.5"
mypy = "^1.5"
sanic-testing = "^22.3.0, <22.9.0"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
python_functions = "test_"
asyncio_mode = "auto"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.3.5,<0.4.0"
pytest-asyncio = "^0.21.0"
